requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0
anthropic>=0.40
google-auth-oauthlib>=1.2
google-api-python-client>=2.150
dropbox>=12.0
